Polishing

+++ rsocket-core/src/main/java/io/rsocket/internal/SynchronizedIntObjectHashMap.java
@@ -189,7 +189,7 @@ public class SynchronizedIntObjectHashMap<V> implements IntObjectMap<V> {
 
   @Override
   public synchronized void clear() {
-    Arrays.fill(keys, (int) 0);
+    Arrays.fill(keys, 0);
     Arrays.fill(values, null);
     size = 0;
   }
@@ -331,7 +331,7 @@ public class SynchronizedIntObjectHashMap<V> implements IntObjectMap<V> {
   }
 
   private int objectToKey(Object key) {
-    return (int) ((Integer) key).intValue();
+    return ((Integer) key).intValue();
   }
 
   /**
@@ -369,7 +369,7 @@ public class SynchronizedIntObjectHashMap<V> implements IntObjectMap<V> {
 
   /** Returns the hash code for the key. */
   private static int hashCode(int key) {
-    return (int) key;
+    return key;
   }
 
   /** Get the next sequential index after {@code index} and wraps if necessary. */
